refactor(desktop): clarify main module dependencies
